speed up program by around 30%
Instead of checking the end of scanning_terminated flag
every time a character is pushed to the would be lexer,
we create a new function that just returns eof, and an fn
pointer in the scanner that initializes with the normal getc.
This way, that overhead of checking is gone, and the compiler can
optimize away the pointer chase problem.