Fix bugs regarding fn ptr and bad internal state
In the main function, we were calling scanner_getc
directly, but we were supposed to call a function that
is a member of scanner. This problem would be fixed if
we had lexa.c include from a header so that the behind the
scenes functions can't be accessed. The other bug was
we weren't switching the next_buff_already_loaded flag
causing the scanner to never read more input.