Implement most of eNFA simulation.
the eNFA_contains function can now consume the
string, generating the new set of states as it goes.
If on the step it sees no other states, it returns false,
with a message, otherwise, it continues processing. It
does not yet do acceptence checking, and does not ever
return true.