Take care of the 0 label case
When the generator got a point that was right on top of the line,
there would be a problem, because it's neither above nor below the
line. I'm not sure, but this could present enough of a problem,
especially when working with floats. So, if a label was ever 0,
we would find another domain point, this seems to fix the infinite
loops we would sometimes notice.